+++ internal/provider/resource_rediscloud_subscription_peering_test.go
@@ -2,18 +2,18 @@ package provider
 
 import (
 	"fmt"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"net"
 	"os"
 	"regexp"
 	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
 func TestAccResourceRedisCloudSubscriptionPeering_aws(t *testing.T) {
 
 	name := acctest.RandomWithPrefix(testResourcePrefix)
-	password := acctest.RandString(20)
 
 	testCloudAccountName := os.Getenv("AWS_TEST_CLOUD_ACCOUNT_NAME")
 
@@ -42,7 +42,6 @@ func TestAccResourceRedisCloudSubscriptionPeering_aws(t *testing.T) {
 		testCloudAccountName,
 		name,
 		subCidrRange,
-		password,
 		peeringRegion,
 		accountId,
 		vpcId,
@@ -79,11 +78,9 @@ func TestAccResourceRedisCloudSubscriptionPeering_gcp(t *testing.T) {
 	}
 
 	name := acctest.RandomWithPrefix(testResourcePrefix)
-	password := acctest.RandString(20)
 
 	tf := fmt.Sprintf(testAccResourceRedisCloudSubscriptionPeeringGCP,
 		name,
-		password,
 		os.Getenv("GCP_VPC_PROJECT"),
 		os.Getenv("GCP_VPC_ID"),
 	)
@@ -163,14 +160,15 @@ resource "rediscloud_subscription" "example" {
     }
   }
 
-  database {
-    name = "tf-database"
-    protocol = "redis"
+  creation_plan {
+    average_item_size_in_bytes = 1
     memory_limit_in_gb = 1
-    data_persistence = "none"
+    quantity = 1
+    replication=false
+    support_oss_cluster_api=false
     throughput_measurement_by = "operations-per-second"
-    password = "%s"
     throughput_measurement_value = 10000
+	modules = []
   }
 }
 
@@ -204,14 +202,15 @@ resource "rediscloud_subscription" "example" {
     }
   }
 
-  database {
-    name = "tf-database"
-    protocol = "redis"
+  creation_plan {
+    average_item_size_in_bytes = 1
     memory_limit_in_gb = 1
-    data_persistence = "none"
+    quantity = 1
+    replication=false
+    support_oss_cluster_api=false
     throughput_measurement_by = "operations-per-second"
-    password = "%s"
     throughput_measurement_value = 10000
+	modules = []
   }
 }